+++ analytics/notebooks/lib/run_meta.py
@@ -18,6 +18,10 @@ class RunMeta():
         self.ts_run_start = to_date(metaJson["meta"]["run_start_time"], perf_meta_date_ts_pattern)
         self.ts_run_end = to_date(metaJson["meta"]["run_end_time"], perf_meta_date_ts_pattern)
         self.metaJson = metaJson
+        self.startPublisherList = self.createStartPublisherList(metaJson)
+        self.endPublisherList = self.createEndPublisherList(metaJson)
+        self.startConsumerList = self.createStartConsumerList(metaJson)
+        self.endConsumerList = self.createEndConsumerList(metaJson)
 
     def __str__(self):
         return f'[RunMeta: [infrastructure: {self.infrastructure}] [run_id: {self.run_id}] [run_duration: {self.run_duration()}]]'
@@ -25,18 +29,128 @@ class RunMeta():
     def run_duration(self):
         return self.ts_run_end - self.ts_run_start
 
+    def getSolacePubSubInfo(self):
+        return self.metaJson["meta"]["solace_pubsub_image"]
+
     def getEnvRegion(self):
         return self.metaJson["meta"]["env"]["region"]
 
     def getEnvZone(self):
         return self.metaJson["meta"]["env"]["zone"]
 
+    """ Client Connections """
+
+    def _createClientList(self, name_list, client_list):
+        _clientList = []
+        for name in name_list:
+            for client in client_list:
+                if client['clientName'] == name['clientName']:
+                    _clientList.append(client)
+        return _clientList         
+
+    def createStartPublisherList(self, metaJson):
+        nameList = metaJson['meta']['client_connections']['start_test']['publisher_list']
+        clientList = metaJson['meta']['client_connections']['start_test']['client_list']
+        return self._createClientList(nameList, clientList)
+
+    def createEndPublisherList(self, metaJson):
+        nameList = metaJson['meta']['client_connections']['end_test']['publisher_list']
+        clientList = metaJson['meta']['client_connections']['end_test']['client_list']
+        return self._createClientList(nameList, clientList)
+
+    def createStartConsumerList(self, metaJson):
+        nameList = metaJson['meta']['client_connections']['start_test']['consumer_list']
+        clientList = metaJson['meta']['client_connections']['start_test']['client_list']
+        return self._createClientList(nameList, clientList)
+
+    def createEndConsumerList(self, metaJson):
+        nameList = metaJson['meta']['client_connections']['end_test']['consumer_list']
+        clientList = metaJson['meta']['client_connections']['end_test']['client_list']
+        return self._createClientList(nameList, clientList)
+
     def getNumClientConnectionsAtStart(self):
-        return self.metaJson["meta"]["client_connections"]["start_test"]["number"]
+        return len(self.getStartTestConsumerList()) + len(self.getStartTestPublisherList())
 
     def getNumClientConnectionsAtEnd(self):        
-        return self.metaJson["meta"]["client_connections"]["end_test"]["number"]
+        return len(self.getEndTestConsumerList()) + len(self.getEndTestPublisherList()) 
 
+    def getStartTestPublisherList(self):
+        return self.startPublisherList
+
+    def getEndTestPublisherList(self):
+        return self.endPublisherList
+
+    def getStartTestConsumerList(self):
+        return self.startConsumerList
+
+    def getEndTestConsumerList(self):
+        return self.endConsumerList
+
+    def _getClientListAggregates(self, client_list):
+        aggregates=dict(
+            rxDiscardedMsgCount=0,
+            rxMsgCount=0,
+            txDiscardedMsgCount=0,
+            txMsgCount=0,
+            averageRxMsgRate=0,
+            averageTxMsgRate=0
+        )
+        for client in client_list:
+            aggregates['rxDiscardedMsgCount'] += client['rxDiscardedMsgCount']
+            aggregates['rxMsgCount'] += client['dataRxMsgCount']
+            aggregates['txDiscardedMsgCount'] += client['txDiscardedMsgCount']
+            aggregates['txMsgCount'] += client['dataTxMsgCount']
+            aggregates['averageRxMsgRate'] += client['averageRxMsgRate']
+            aggregates['averageTxMsgRate'] += client['averageTxMsgRate']
+
+        return aggregates    
+
+    def getPublisherAggregates(self):
+        end     = self._getClientListAggregates(self.getEndTestPublisherList())
+        start   = self._getClientListAggregates(self.getStartTestPublisherList())
+        return dict(
+            rxDiscardedMsgCount = end['rxDiscardedMsgCount']   - start['rxDiscardedMsgCount'],
+            rxMsgCount          = end['rxMsgCount']            - start['rxMsgCount'],
+            txDiscardedMsgCount = end['txDiscardedMsgCount']   - start['txDiscardedMsgCount'],
+            txMsgCount          = end['txMsgCount']            - start['txMsgCount'],
+            meanRxMsgRate       = end['averageRxMsgRate'],
+            meanTxMsgRate       = end['averageTxMsgRate']
+        )
+
+    def getConsumerAggregates(self):
+        end     = self._getClientListAggregates(self.getEndTestConsumerList())
+        start   = self._getClientListAggregates(self.getStartTestConsumerList())
+        return dict(
+            rxDiscardedMsgCount = end['rxDiscardedMsgCount'] - start['rxDiscardedMsgCount'],
+            rxMsgCount          = end['rxMsgCount'] - start['rxMsgCount'],
+            txDiscardedMsgCount = end['txDiscardedMsgCount'] - start['txDiscardedMsgCount'],
+            txMsgCount          = end['txMsgCount'] - start['txMsgCount'],
+            meanRxMsgRate       = end['averageRxMsgRate'],
+            meanTxMsgRate       = end['averageTxMsgRate']
+        )
+
+    def getConsumerNamesValues4Plotting(self):
+        start_list = self.getStartTestConsumerList()
+        end_list   = self.getEndTestConsumerList()
+        names = []
+        values = []
+        for consumer_start, consumer_end  in zip(start_list, end_list):
+            name=consumer_end['clientName']
+            # NOTE: very brittle, assumes the following structure: 
+            # sdkperf-load@devel1-consumer-node-1-consumer_2-000001
+            # go from the back - user defines infrastructure name
+            parts = name.split("-")
+            import logging, sys
+            logging.debug(f"name={name}")
+            logging.debug(f"parts = {parts}")
+            node_number = parts[len(parts)-1-2]
+            consumer_name = parts[len(parts)-1-1]
+            logging.debug(f"node_number={node_number}, consumer_name={consumer_name}")
+            names.append(f"node-{node_number}:{consumer_name}")
+            val = consumer_end['dataTxMsgCount'] - consumer_start['dataTxMsgCount']
+            values.append(val)
+        return names, values
+    
     def getRunSpecDescription(self):
         return self.metaJson["meta"]["run_spec"]["general"]["description"]
     
@@ -85,59 +199,45 @@ class RunMeta():
             return "n/a"    
         return len(self.metaJson["meta"]["run_spec"]["load"]["subscribe"]["consumers"])
 
+    def getRunSpecLoadSubscribeDistributionStrategy(self):
+        if not self.getRuncSpecLoadIsIncluded():
+            return "n/a"    
+        return self.metaJson["meta"]["run_spec"]["load"]["subscribe"]["consumer_distribution_strategy"]
+
+    def getNodeSpec(self, node): 
+        if self.cloud_provider == "azure":
+            return f"size: {node['size']}"
+        elif self.cloud_provider == "aws":
+            return f"type: {node['node_details']['instance_type']}, cores: {node['node_details']['cpu_core_count']}"
+        else:    
+            return f"ERROR: unknown node spec for cloud_provider:{self.cloud_provider}"
+
     def getNumBrokerNodes(self):
         return len(self.metaJson["meta"]["nodes"]["broker_nodes"])
 
-    def getBrokerNodeDetails(self, public_ip):
-        broker_nodes = self.metaJson["meta"]["nodes"]["broker_nodes"]
-        # there can only be 1
-        if len(broker_nodes) != 1:
-            raise PerfError("[ERROR] - found more than 1 broker_node in nodes")
-        if broker_nodes[0]["public_ip"] != public_ip:
-            raise PerfError("[ERROR] - broker node public ip does not match inventory public_ip")
-        return broker_nodes[0]
-
     def getBrokerNodeSpec(self):
-        public_ip = self.metaJson["meta"]["inventory"]["all"]["hosts"]["broker_centos"]["ansible_host"]
-        node = self.getBrokerNodeDetails(public_ip)
-        if self.cloud_provider == "azure":
-            return f"size: {node['size']}"
-        elif self.cloud_provider == "aws":
-            return f"type: {node['node_details']['instance_type']}, cores: {node['node_details']['cpu_core_count']}"
-        else:    
-            return f"ERROR: unknown node spec for cloud_provider:{self.cloud_provider}"
-
-    def getMonitorNodeDetails(self, public_ip):
-        sdkperf_nodes = self.metaJson["meta"]["nodes"]["sdkperf_nodes"]
-        return [x for x in sdkperf_nodes if x["public_ip"] == public_ip][0]
+        node = self.metaJson["meta"]["nodes"]["broker_nodes"][0]
+        return self.getNodeSpec(node)
 
     def getMonitorNodeSpec(self):
-        sdkperf_latency_hosts = self.metaJson["meta"]["inventory"]["sdkperf_latency"]["hosts"]
-        # assuming here there is only 1
-        if len(sdkperf_latency_hosts) != 1:
-            raise PerfError("[ERROR] - found more than 1 sdkperf_latency host in inventory")
-        monitor_node_public_ip = sdkperf_latency_hosts[list(sdkperf_latency_hosts)[0]]["ansible_host"]
-        node = self.getMonitorNodeDetails(monitor_node_public_ip)
-        if self.cloud_provider == "azure":
-            return f"size: {node['size']}"
-        elif self.cloud_provider == "aws":
-            return f"type: {node['node_details']['instance_type']}, cores: {node['node_details']['cpu_core_count']}"
-        else:    
-            return f"ERROR: unknown node spec for cloud_provider:{self.cloud_provider}"
-
-    def getPublisherNodeSpec(self):
-        # lazy, assuming they are the same
-        return self.getMonitorNodeSpec()
+        node = self.metaJson["meta"]["nodes"]["latency_nodes"][0]
+        return self.getNodeSpec(node)
 
     def getNumPublisherNodes(self):
-        return len(self.metaJson["meta"]["inventory"]["sdkperf_publishers"]["hosts"])
+        #  -1:NOT_A_HOST
+        return len(self.metaJson["meta"]["inventory"]["sdkperf_publishers"]["hosts"])-1
 
-    def getConsumerNodeSpec(self):
-        # lazy, assuming they are the same
-        return self.getMonitorNodeSpec()
+    def getPublisherNodeSpec(self):
+        node = self.metaJson["meta"]["nodes"]["publisher_nodes"][0]
+        return self.getNodeSpec(node)
 
     def getNumConsumerNodes(self):
-        return len(self.metaJson["meta"]["inventory"]["sdkperf_consumers"]["hosts"])
+        #  -1:NOT_A_HOST
+        return len(self.metaJson["meta"]["inventory"]["sdkperf_consumers"]["hosts"])-1
+
+    def getConsumerNodeSpec(self):
+        node = self.metaJson["meta"]["nodes"]["consumer_nodes"][0]
+        return self.getNodeSpec(node)
 
     """ Run Spec General """
     def getRunSpecGeneral(self):
@@ -178,9 +278,23 @@ class RunMeta():
     def getRuncSpecMonitorVpnIsIncluded(self):
         return self.metaJson["meta"]["run_spec"]["monitors"]["vpn_stats"]["include"]
 
+    def getDisplayNameCloudProvider(self):
+        if self.cloud_provider == "azure": 
+            return "Azure"
+        elif self.cloud_provider == "aws":
+            return "AWS"
+        else:    
+            return self.cloud_provider
+
     def getUseCaseAsMarkdown(self):
         md = f"""
-## Use Case: {self.metaJson["meta"]["use_case"]}
+## Use Case: {self.metaJson["meta"]["run_spec"]["general"]["use_case"]["display_name"]} ({self.metaJson["meta"]["run_spec"]["general"]["use_case"]["name"]})
+
+Test Specification: 
+{self.getRunSpecGeneral()["test_spec"]["descr"]}.
+({self.getRunSpecGeneral()["test_spec"]["name"]})
+
+Cloud Provider: **{self.getDisplayNameCloudProvider()}**
         """
         return md
 
@@ -190,16 +304,15 @@ class RunMeta():
 
 ## Run Settings
 * Description: "{self.getRunSpecDescription()}"
-* test spec id: ({self.getRunSpecGeneral()["test_spec_name"]})
 
 |General                    |                                                   | | Infrastructure:           | cloud provider:{self.cloud_provider}                                        |      |   |  
 |:--------------------------|:--------------------------------------------------|-|:--------------------------|:----------------------------------------------------------------------------|:-----|:--|
 |Run name:                  |{self.run_name}                                    | |{self.infrastructure}      |region: {self.getEnvRegion()}, zone: {self.getEnvZone()}                     |      |   |                                              
-|Run Id:                    |{self.run_id}                                      | |Broker Node:               |nodes: {self.getNumBrokerNodes()}<br/>spec: {self.getBrokerNodeSpec()}       |      |   |
-|Run Start:                 |{self.ts_run_start}                                | |Load<br/>Publisher Nodes:  |nodes: {self.getNumPublisherNodes()}<br/>spec:{self.getPublisherNodeSpec()}  |      |   |
-|Run End:                   |{self.ts_run_end}                                  | |Load<br/>Consumer Nodes:   |nodes: {self.getNumConsumerNodes()} <br/>spec:{self.getConsumerNodeSpec()}   |      |   |
-|Run Duration:              |{self.run_duration()}                              | |Monitor Node:              |nodes: {self.getNumMonitorNodes()} <br/>spec:{self.getMonitorNodeSpec()}     |      |   |  
-|Sample Duration (secs):    |{self.getRunSpecParamsSampleDurationSecs()}        | |                           |                                                                             |      |   | 
+|Run Id:                    |{self.run_id}                                      | |Broker Node:               |nodes: {self.getNumBrokerNodes()}<br/>{self.getBrokerNodeSpec()}             |      |   |
+|Run Start:                 |{self.ts_run_start}                                | |Load<br/>Publisher Nodes:  |nodes: {self.getNumPublisherNodes()}<br/>{self.getPublisherNodeSpec()}       |      |   |
+|Run End:                   |{self.ts_run_end}                                  | |Load<br/>Consumer Nodes:   |nodes: {self.getNumConsumerNodes()} <br/>{self.getConsumerNodeSpec()}        |      |   |
+|Run Duration:              |{self.run_duration()}                              | |Monitor Node:              |nodes: {self.getNumMonitorNodes()} <br/>{self.getMonitorNodeSpec()}          |      |   |  
+|Sample Duration (secs):    |{self.getRunSpecParamsSampleDurationSecs()}        | |Solace PubSub+             | {self.getSolacePubSubInfo()}                                                |      |   | 
 |Number of Samples:         |{self.getRunSpecParamsTotalNumSamples()}           | |                           |                                                                             |      |   | 
 
 
@@ -213,6 +326,7 @@ class RunMeta():
 | - Rate (1/sec):       |{self.getRunSpecLoadPublishTotalMsgRatePerSec()}       | |**Ping**       | included:                 |**{self.getRuncSpecMonitorPingIsIncluded()}**                    |  
 | - Topics:             |{self.getRunSpecLoadPublishTotalNumberOfTopics()}      | |**Broker VPN** | included:                 |**{self.getRuncSpecMonitorVpnIsIncluded()}**                     |  
 |Consumers:             |{self.getRunSpecLoadSubscribeTotalNumberOfConsumers()} | |               |                           |                                                                 |   
+| - deploy strategy:    |{self.getRunSpecLoadSubscribeDistributionStrategy()}   | |               |                           |                                                                 |   
 
             """
         return md