+++ uc-non-persistent/tests/auto-run/abort.tp-002.sh
@@ -6,23 +6,21 @@
 
 scriptDir=$(cd $(dirname "$0") && pwd);
 scriptName=$(basename $(test -L "$0" && readlink "$0" || echo "$0"));
+projectHome=${scriptDir%/uc-non-persistent/*}
+usecaseHome=$projectHome/uc-non-persistent
+export TMP_DIR="$scriptDir/../tmp"
+
+export TEST_SPEC_FILE="$scriptDir/tp-002.test.spec.yml"
+export TEST_SPEC_INVENTORY_FILE="$TMP_DIR/test-specs/auto_tp_002.test.spec.inventory.yml"
 
 
-#  format: {cloud_provider}.{config}
-export infrastructureIds=(
-  "azure.test1"
-  # "azure.test2"
-  "aws.test1"
-)
+export SHARED_SETUP_DIR=$usecaseHome/shared-setup;
 
-export INFRASTRUCTURE_IDS="${infrastructureIds[*]}"
+export ANSIBLE_VERBOSITY=0
+nohup ../_abort.sh > ./logs/$scriptName.out 2>&1 &
 
-export LOG_DIR=$scriptDir/logs
-rm -f $LOG_DIR/**destroy**
-
-export TF_VARIABLES_DIR=$scriptDir
-
-nohup ../_run.destroy-all.sh > $LOG_DIR/$scriptName.out 2>&1 &
+# export ANSIBLE_VERBOSITY=3
+# ../_abort.sh
 
 ###
 # The End.