+++ uc-non-persistent/tests/auto-run/run.tp-002.bg.sh
@@ -6,23 +6,14 @@
 
 scriptDir=$(cd $(dirname "$0") && pwd);
 scriptName=$(basename $(test -L "$0" && readlink "$0" || echo "$0"));
+projectHome=${scriptDir%/uc-non-persistent/*}
+usecaseHome=$projectHome/uc-non-persistent
 
+export TEST_SPEC_FILE="$scriptDir/tp-002.test.spec.yml"
 
-#  format: {cloud_provider}.{config}
-export infrastructureIds=(
-  "azure.test1"
-  # "azure.test2"
-  "aws.test1"
-)
+export ANSIBLE_VERBOSITY=0
 
-export INFRASTRUCTURE_IDS="${infrastructureIds[*]}"
-
-export LOG_DIR=$scriptDir/logs
-rm -f $LOG_DIR/**destroy**
-
-export TF_VARIABLES_DIR=$scriptDir
-
-nohup ../_run.destroy-all.sh > $LOG_DIR/$scriptName.out 2>&1 &
+nohup ../_run.sh > ./logs/$scriptName.out 2>&1 &
 
 ###
 # The End.